feat: change menu and exit menu buttons
Two small changes:

- Switch "menu" button to a hamburger menu.
- Switch the "exit menu" button in the menu to an X; IMO using a
  hamburger menu to mean "close the menu" is a bit confusing. I also
  made the X button show up on mobile as it might be confusing how to
  close the menu without it.

Question - should we hide the menu button if the menu is open? Not hard
to do but just curious if we would want that behaviour. Only thing is
that it going invisible/hidden might look a bit weird due to the
chapter/page text right next to it, so that might need to also be looked
at.