feat: disable language selection if there's only one chapter language
Admittedly this probably isn't needed once we get all the translations
done, but something I noticed right now is that our code _really_
doesn't like it if you try to switch to JP since we only have English
chapters, and errors everywhere. So I guess as a (temporary) safety
measure, just prevent users from switching the language if there's no
language to switch to in the first place.

Not sure if we would prefer to just disable the button or hide it
altogether in this case; probably doesn't matter though as we likely
won't have to worry about this when we actually release it. I hope.