fix: fix missing progress bar on mobile due to toolbar height
The progress bar was hidden due to the variable height of phone
browsers from the toolbar. This was because we set the reader
height to be `100vh`; this changes it to be `100dvh` which
avoids the problem.

_Might_ be better to push the h-dvh to the reader itself, but I
think it's better like this (or more like it's less likely to
cause problems).